# shared/openapi/openapi.yaml
openapi: 3.0.2
paths:
...
  /auth/login-requests:
    post:
      tags: &a2
        - auth
      parameters: []
      operationId: auth.loginRequest
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                friendCode:
                  type: string
                  minLength: 1
                skipUpdateScore:
                  default: true
                  type: boolean
                method:
                  type: string
                  enum:
                    - bot_sends_request
                    - user_sends_request
              required:
                - friendCode
                - method
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  authUrl:
                    type: string
                  authToken:
                    type: string
                  message:
                    type: string
                  reused:
                    type: boolean
                  botFriendCode:
                    type: string
                  createdAt:
                    type: string
                  job:
                    type: object
                    properties:
                      id:
                        type: string
                      friendCode:
                        type: string
                      jobType:
                        type: string
                        enum:
                          - send_friend_request
                          - accept_friend_request
                          - update_score
                          - get_user_recent_event
                      priority:
                        type: integer
                      skipUpdateScore:
                        type: boolean
                      botUserFriendCode:
                        type: string
                        nullable: true
                      friendRequestSentAt:
                        type: string
                        nullable: true
                      friendRequestWaitStartedAt:
                        type: string
                        nullable: true
                      status:
                        type: string
                        enum:
                          - queued
                          - processing
                          - completed
                          - failed
                          - canceled
                      stage:
                        type: string
                        enum:
                          - send_request
                          - wait_acceptance
                          - wait_user_request
                          - accept_request
                          - update_score
                          - get_user_recent_event
                      profile:
                        nullable: true
                      error:
                        type: string
                        nullable: true
                      executing:
                        type: boolean
                      scoreProgress:
                        type: object
                        properties:
                          completedDiffs:
                            type: array
                            items:
                              type: integer
                              minimum: 0
                              maximum: 14
                          totalDiffs:
                            type: integer
                            minimum: 0
                        required:
                          - completedDiffs
                          - totalDiffs
                        nullable: true
                      updateScoreDuration:
                        type: number
                        nullable: true
                      autoExportResult:
                        type: object
                        properties:
                          divingFish:
                            type: object
                            properties:
                              status:
                                type: string
                              message:
                                type: string
                            required:
                              - status
                            nullable: true
                          lxns:
                            type: object
                            properties:
                              status:
                                type: string
                              message:
                                type: string
                            required:
                              - status
                            nullable: true
                        nullable: true
                      isAuthenticated:
                        type: boolean
                      cabinetScoreMap:
                        type: object
                        additionalProperties:
                          type: object
                          properties:
                            achievement:
                              type: number
                            dxScore:
                              type: number
                          required:
                            - achievement
                            - dxScore
                        nullable: true
                      diffsToScrape:
                        type: array
                        items:
                          type: integer
                        nullable: true
                      runAt:
                        type: string
                        nullable: true
                      createdAt:
                        type: string
                      updatedAt:
                        type: string
                    required:
                      - id
                      - friendCode
                      - jobType
                      - skipUpdateScore
                      - status
                      - stage
                      - createdAt
                      - updatedAt
                required:
                  - jobId
                additionalProperties: true
  /auth/login-requests/{jobId}:
    get:
      tags: *a2
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      operationId: auth.loginStatus
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  done:
                    type: boolean
                  status:
                    type: string
                  friendCode:
                    type: string
                  message:
                    type: string
                additionalProperties: true
  /auth/login-requests/{jobId}/verify:
    post:
      tags: *a2
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      operationId: auth.verifyLoginRequest
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  job:
                    type: object
                    properties:
                      id:
                        type: string
                      friendCode:
                        type: string
                      jobType:
                        type: string
                        enum:
                          - send_friend_request
                          - accept_friend_request
                          - update_score
                          - get_user_recent_event
                      priority:
                        type: integer
                      skipUpdateScore:
                        type: boolean
                      botUserFriendCode:
                        type: string
                        nullable: true
                      friendRequestSentAt:
                        type: string
                        nullable: true
                      friendRequestWaitStartedAt:
                        type: string
                        nullable: true
                      status:
                        type: string
                        enum:
                          - queued
                          - processing
                          - completed
                          - failed
                          - canceled
                      stage:
                        type: string
                        enum:
                          - send_request
                          - wait_acceptance
                          - wait_user_request
                          - accept_request
                          - update_score
                          - get_user_recent_event
                      profile:
                        nullable: true
                      error:
                        type: string
                        nullable: true
                      executing:
                        type: boolean
                      scoreProgress:
                        type: object
                        properties:
                          completedDiffs:
                            type: array
                            items:
                              type: integer
                              minimum: 0
                              maximum: 14
                          totalDiffs:
                            type: integer
                            minimum: 0
                        required:
                          - completedDiffs
                          - totalDiffs
                        nullable: true
                      updateScoreDuration:
                        type: number
                        nullable: true
                      autoExportResult:
                        type: object
                        properties:
                          divingFish:
                            type: object
                            properties:
                              status:
                                type: string
                              message:
                                type: string
                            required:
                              - status
                            nullable: true
                          lxns:
                            type: object
                            properties:
                              status:
                                type: string
                              message:
                                type: string
                            required:
                              - status
                            nullable: true
                        nullable: true
                      isAuthenticated:
                        type: boolean
                      cabinetScoreMap:
                        type: object
                        additionalProperties:
                          type: object
                          properties:
                            achievement:
                              type: number
                            dxScore:
                              type: number
                          required:
                            - achievement
                            - dxScore
                        nullable: true
                      diffsToScrape:
                        type: array
                        items:
                          type: integer
                        nullable: true
                      runAt:
                        type: string
                        nullable: true
                      createdAt:
                        type: string
                      updatedAt:
                        type: string
                    required:
                      - id
                      - friendCode
                      - jobType
                      - skipUpdateScore
                      - status
                      - stage
                      - createdAt
                      - updatedAt
                required:
                  - job
  /auth/qr-login:
    post:
      tags: *a2
      parameters: []
      operationId: auth.loginByQr
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                qrCode:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      friendCode:
                        type: string
                    required:
                      - friendCode
                    additionalProperties: true
                required:
                  - token
                  - user
        "400":
          description: "400"
        "404":
          description: "404"
        "409":
          description: "409"
  /auth/qr-login/{attemptId}:
    get:
      tags: *a2
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: string
      operationId: auth.pollLoginByQr
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                nullable: true
        "400":
          description: "400"
...
info:
  title: Maimai Score Hub API
  version: 1.0.0
  description: Generated from @maimai-score-hub/shared ts-rest contracts
servers:
  - url: /api/v1